python/requirements.txt
langchain>=0.1.0
langgraph>=0.0.15
anthropic>=0.17.0
python-dotenv>=0.19.0
langchain-anthropic>=0.0.1
